pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sea-analytics"
version = "0.1.0"
description = "Stateful Execution Environment for Conversational Big Data Analytics (NeurIPS 2025)"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    # Core framework
    "pydantic>=2.0",

    # LLM / Planner
    "google-genai>=1.0",
    "langchain-google-genai>=2.0",
    "langchain>=0.3",
    "langchain-community>=0.3",

    # Vector store (SchemaRetriever)
    "chromadb>=0.5",
    "sentence-transformers>=3.0",

    # Data access
    "pandas>=2.0",
    "numpy>=1.26",
    "databricks-sql-connector>=3.0",
    "boto3>=1.34",           # S3 data backend

    # Visualisation (ChartGenerator)
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.0",
    "Pillow>=10.0",

    # API server (optional but included so `sea.api` is always importable)
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",      # FastAPI test client
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sea*"]

[tool.setuptools.package-data]
"sea" = ["py.typed"]
